co-locate @glimmer/component docs with the actual implementation
